Fix incorrect data sent during Game Boy Player handshake
This commit fixes an out-of-bounds buffer read (due to an off by one
error) that leads to the library sending the wrong response to the Game
Boy Player immediately after the "NINTENDO" handshake.

After 'B0BB4F44', the GBA should send back '8000B0BB'. However, the code
was actually reading past the end of the "NINTENDO" array, sending
something like 'FF00B0BB' (depending on how the code is compiled).

Given that this apparently worked before, the Game Boy Player evidently
doesn't verify that (or there are multiple correct answers). However,
all other games I tested properly send '8000', and this is clearly the
intention of the code.